+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Upcoming
+
+### Changed
+
+- [__BREAKING__] Generated bindings via tree-sitter-cli v0.23.0  
+  With tree-sitter v0.23.0, Rust bindings have changed.
+
 
 ## Version 1.0.0 - 2024-06-25